refactor(geometry): split conic edge samplers into advanced_face/conics.rs
curves.rs crossed the 400-line new-file ratchet threshold with the trim
fix; the axis-placement reader and circle/ellipse arc samplers move to
their own module. No behavior change.
